Refactor bpjs services
